Fixed helicopter on multiplayer servers ( #409 ) (#414)
* Fixed helicopter on multiplayer servers ( #409 )

* Updated version number

* Improved the code a bit

The helicopter will now spawn with the player's yaw rotation + Slimmed
the code

* Reverted changelog file

- I accidentally deleted my previous change from the changelog

+++ src/main/java/org/jurassicraft/server/entity/vehicle/FordExplorerEntity.java
@@ -146,7 +146,7 @@ public class FordExplorerEntity extends CarEntity {
         if(onRails) {
             if(this.canPassengerSteer()) {
                 if (this.getPassengers().isEmpty() || !(this.getPassengers().get(0) instanceof EntityPlayer)) {
-                    this.setControlState(0);
+                    this.setControlState((byte) 0);
                 }
                 if(this.world.isRemote) {
                     this.handleControl(); //+Z-X
